feat(input): add ux-input component (#34)
* feat(ux-input): Add ux-input core to Aurelia UX

Added ux-input files and setup core implimentation files. Added ux-input to index.ts.

* feat(ux-input): Add full-width type

* feat(ux-input): Add binding to value

Binds to ux-input value. Defaults to two way databinding.

* style(ux-input): Edited the border to cover sides

* feat(ux-input): Add input counter

* style(ux-input): Color counter accent on select

* feat(ux-input): Added disable attribute

* feat(ux-input): Add placeholder

Add placeholder attribute
Hookup blur event

* feat(ux-input): Add placeholder

Add placeholder attribute
Hookup blur event

* feat(ux-input): Add standard attributes

Added:
Readonly,
Step

* feat(ux-input): Add standard attributes

Added:
Readonly,
Step

* refactor(ux-input): Reorganise bindings

* refactor(ux-input): Reorganise bindings

* feat(ux-input): Add autofocus attribute

Autofocus on the ux-input element fires a focus on the contained input element. This allows the autofocus to happen more than once.

* feat(ux-input): Add autofocus attribute

Autofocus on the ux-input element fires a focus on the contained input element. This allows the autofocus to happen more than once.

* feat(ux-input): Add error class

Add CSS Class to display errors. This causes the control to gain a red border when the class is applied.

* feat(ux-input): Add error class

Add CSS Class to display errors. This causes the control to gain a red border when the class is applied.

* refactor(ux-input): Change attribute binding

Changed the method in which attributes are bound.

* refactor(ux-input): Change attribute binding

Changed the method in which attributes are bound.

* refactor(ux-input): Remove unused theme properties

* refactor(ux-input): Remove unused theme properties

* feat(ux-input): Add ux-input-info

UX-Input-Info is a property for handling showing messages and showing a ux-input-counter for an input element.

* feat(ux-input): Add ux-input-info

UX-Input-Info is a property for handling showing messages and showing a ux-input-counter for an input element.

* refactor(ux-input): Moved ux-input-info styles

* refactor(ux-input): Moved ux-input-info styles

* feat(ux-input): Counter added to ux-input-info

* feat(ux-input): Counter added to ux-input-info

* feat(ux-input): Added hint text slot

* feat(ux-input): Added hint text slot

* feat(ux-input): Add focused class on input focus

When the input of UX-Input gains focus, it adds 'focused' to the class list. When the blur event happens, it removes 'focused' from the class list.

* feat(ux-input): Add focused class on input focus

When the input of UX-Input gains focus, it adds 'focused' to the class list. When the blur event happens, it removes 'focused' from the class list.

* style(ux-input): Changed left and right padding

Change left padding on full-width input to 8px
Change right padding on full-width input to 8px

* style(ux-input): Changed left and right padding

Change left padding on full-width input to 8px
Change right padding on full-width input to 8px

* style(ux-input): Filled out themes

Code cleanup, preparing for pull reqeust.
Remove required.
Limit type to text and number.

* style(ux-input): Filled out themes

Code cleanup, preparing for pull reqeust.
Remove required.
Limit type to text and number.

* feat(ux-input): Added disable attribute

* feat(input): Check for valid number input

Add Min attribute
Add Max attribute
Check that input is a valid number, not below the minimum, and not above the maximum.

* refactor(ux-input): Remove test code