fix(grid): grid based on grid element size
The previous behavior was to determine the size (xs, sm, md, lg, xl)
based on the window width. This is fine when the grid is placed in full width
but doesn't work well when it's placed inside an already constraint element.

This commit suggests to change this behavior and base the grid reparation
on the actual grid element width.